Fix and extend unit tests.
This patch fixes and extends the code and unit tests provided by the
Geotrellis team.  There are three tests of the histogram-calculating:
one polygon that covers the entire extent, one that is a strict subset
of the extent, and one that misses the extent.

In the first case, each value in the histogram is expected to be equal
to those of a histogram calculated in the naive fashion.  In the second
case, each value of the histogram is expected to be less than or equal
to those of the histogram computed over the entire extent.  In the third
case, an empty histogram is expected.